third_party/blink/web_tests/fast/css/min-device-aspect-ratio-expected.txt

This text is green if the min-device-aspect-ratio media query works properly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS successfullyParsed is true

TEST COMPLETE

PASS document.defaultView.getComputedStyle(document.getElementById('description'), null).color is 'rgb(0, 128, 0)'
PASS successfullyParsed is true

TEST COMPLETE
